refactor(test): share state index-drift fixture
